code_generator/requirements.txt
requests==2.32.3
openai==1.51.0
pyyaml==6.0.2
